refactor: extract IBasalSeriesBuilder from IobCobComputeStage
Move basal series construction (BuildBasalSeriesFromTempBasalsAsync,
BuildBasalSeriesFromProfileAsync, MapTempBasalOrigin) into a standalone
IBasalSeriesBuilder/BasalSeriesBuilder service so a lightweight reports
endpoint can build basal series without triggering the full IOB/COB
compute pipeline.
